Reworking how Errors are handled
Added a custom Error for not OK (200 code) responses. These inherit from Error, but also allow for accessing the `response` and `bodyString` as was the original intention of the code.
Without this custom error, one would receive back: 'Error: [object Object]' when logging or attempting to use the error returned.

Also updated the no element found for explicitly required Error such that it will return with the message intended instead of with that message being hidden by an internal `toString` conversion within Error.